HTBHF-6 restructure apply journey directory
