+++ src/web/routes/application/are-you-pregnant/test/validation-composition.test.js
@@ -3,7 +3,7 @@ const { assocPath } = require('ramda')
 const { applyExpressValidation } = require('../../common/test/apply-express-validation')
 const { dateAsString } = require('../../common/formatters')
 const { validate } = require('../validate')
-const { YES } = require('../../constants')
+const { YES } = require('../../common/constants')
 
 const translate = string => string
 const sixMonthsInFuture = dateAsString({ monthAdjustment: 6 }).split('-')
