+++ src/web/routes/application/common/constants.js
