+++ src/web/routes/application/middleware/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+  ...require('./configure-post'),
+  ...require('./handle-post'),
+  ...require('./render-view'),
+  ...require('./sanitize'),
+  ...require('./session-details')
+}
