+++ src/web/routes/application/middleware/render-view/index.js
