+++ src/web/routes/application/middleware/render-view/render-view.js
@@ -1,4 +1,4 @@
-const { stateMachine, actions } = require('../state-machine')
+const { stateMachine, actions } = require('../../common/state-machine')
 
 const renderView = (template, getPageContent, redirect) => (req, res) => {
   if (req.method === 'POST' && !res.locals.errors) {
