+++ src/web/routes/application/middleware/session-details.js
